# perf: use msm gpu when generating groth16 proof