--- a/tachyon/math/elliptic_curves/msm/algorithms/icicle/BUILD.bazel
+++ b/tachyon/math/elliptic_curves/msm/algorithms/icicle/BUILD.bazel
@@ -17,6 +17,7 @@ tachyon_cuda_library(
     hdrs = ["icicle_msm.h"],
     deps = [
         ":icicle_msm_bn254",
+        "//tachyon/base:bit_cast",
         "//tachyon/device/gpu:gpu_device_functions",
         "//tachyon/device/gpu:gpu_enums",
         "//tachyon/math/elliptic_curves:points",