--- a/tachyon/math/elliptic_curves/msm/variable_base_msm_gpu_unittest.cc
+++ b/tachyon/math/elliptic_curves/msm/variable_base_msm_gpu_unittest.cc
@@ -7,7 +7,6 @@
 #include "tachyon/device/gpu/gpu_enums.h"
 #include "tachyon/device/gpu/scoped_mem_pool.h"
 #include "tachyon/device/gpu/scoped_stream.h"
-#include "tachyon/math/elliptic_curves/bn/bn254/g1.h"
 #include "tachyon/math/elliptic_curves/msm/test/variable_base_msm_test_set.h"
 
 namespace tachyon::math {
@@ -18,15 +17,18 @@ constexpr size_t kThreadNum = 32;
 
 using namespace device;
 
+template <typename Point>
 class VariableMSMCorrectnessGpuTest : public testing::Test {
  public:
+  using Curve = typename Point::Curve;
+
   constexpr static size_t kLogCount = 10;
   constexpr static size_t kCount = 1 << kLogCount;
 
   static void SetUpTestSuite() {
-    bn254::G1Curve::Init();
+    Point::Curve::Init();
 
-    test_set_ = VariableBaseMSMTestSet<bn254::G1AffinePoint>::Random(
+    test_set_ = VariableBaseMSMTestSet<Point>::Random(
         kCount, VariableBaseMSMMethod::kMSM);
 
     expected_ = test_set_.answer.ToProjective();
@@ -35,17 +37,25 @@ class VariableMSMCorrectnessGpuTest : public testing::Test {
   static void TearDownTestSuite() { GPU_MUST_SUCCESS(gpuDeviceReset(), ""); }
 
  protected:
-  static VariableBaseMSMTestSet<bn254::G1AffinePoint> test_set_;
-  static bn254::G1ProjectivePoint expected_;
+  static VariableBaseMSMTestSet<Point> test_set_;
+  static ProjectivePoint<Curve> expected_;
 };
 
-VariableBaseMSMTestSet<bn254::G1AffinePoint>
-    VariableMSMCorrectnessGpuTest::test_set_;
-bn254::G1ProjectivePoint VariableMSMCorrectnessGpuTest::expected_;
+template <typename Point>
+VariableBaseMSMTestSet<Point> VariableMSMCorrectnessGpuTest<Point>::test_set_;
+template <typename Point>
+ProjectivePoint<typename Point::Curve>
+    VariableMSMCorrectnessGpuTest<Point>::expected_;
 
 }  // namespace
 
-TEST_F(VariableMSMCorrectnessGpuTest, MSM) {
+using PointTypes = testing::Types<bn254::G1AffinePoint, bn254::G2AffinePoint>;
+TYPED_TEST_SUITE(VariableMSMCorrectnessGpuTest, PointTypes);
+
+TYPED_TEST(VariableMSMCorrectnessGpuTest, MSM) {
+  using Point = TypeParam;
+  using Curve = typename Point::Curve;
+
   gpuMemPoolProps props = {gpuMemAllocationTypePinned,
                            gpuMemHandleTypeNone,
                            {gpuMemLocationTypeDevice, 0}};
@@ -58,11 +68,11 @@ TEST_F(VariableMSMCorrectnessGpuTest, MSM) {
 
   gpu::ScopedStream stream = gpu::CreateStream();
 
-  VariableBaseMSMGpu<bn254::G1AffinePoint> msm_gpu(mem_pool.get(),
-                                                   stream.get());
-  bn254::G1ProjectivePoint actual;
-  ASSERT_TRUE(msm_gpu.Run(test_set_.bases, test_set_.scalars, &actual));
-  EXPECT_EQ(actual, expected_);
+  VariableBaseMSMGpu<Point> msm_gpu(mem_pool.get(), stream.get());
+  ProjectivePoint<Curve> actual;
+  ASSERT_TRUE(
+      msm_gpu.Run(this->test_set_.bases, this->test_set_.scalars, &actual));
+  EXPECT_EQ(actual, this->expected_);
 }
 
 }  // namespace tachyon::math