… tachyon/math/elliptic_curves/msm/algorithms/icicle/icicle_msm_bn254.cc
… tachyon/math/elliptic_curves/msm/algorithms/icicle/icicle_msm_bn254.h
… third_party/icicle/icicle.BUILD
… vendors/circom/README.md