# Description

This PR uses msm gpu when generating groth16 proofs.

The benchmarking result based on circom proof generation received from Orbiter Finance Team are as follows:

```
# Gpu
Run 1, Time Taken: {Parse ZKey: 29.472 s, Parse Wtns: 0.505613 s, Prove: 14.5769 s, Verify: 0.005145 s} 
# Cpu
Run 1, Time Taken: {Parse ZKey: 29.4742 s, Parse Wtns: 0.498701 s, Prove: 32.4379 s, Verify: 0.001405 s}
```