@chokobole 2024-07-08
> LGTM. Please update the "Speed" section in circom README after GPU benchmark.

Since MSM Gpu doesn't improve the performance on `sha256_512.circom`, we need to find additional circom file to benchmark to show performance gain. Maybe https://github.com/kroma-network/tachyon/tree/benchmark/plonky2-circom is one candidate, the problem is the size of the files are too large..